+++ ChangeLog.md
@@ -7,11 +7,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 
 
-## [1.5.2] - 2019-05-xx
+## [1.6.0] - 2019-05-xx
 
 ### added
 
-- added ...
+- added functions `regex/find-group`, `regex/find-all-groups`
+
+### incompatible changes
+
+- renamed function `match` to `match?` (`match` did not follow the naming conventions)
+- removed function `match-not` to `match-not?` (`match-not` did not follow the naming conventions)
 
 
 
@@ -45,7 +50,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - fixed dereferencing a `future` when the future has thrown an exception
 - fixed `io/copy-stream`
 
-### incompatible changesi
+### incompatible changes
 
 - removed 'io/spit-temp-file', replace with 'io/spit'
 - removed 'io/slurp-temp-file', replace with 'io/slurp'